Revert "Use smaller integer columns for anime and episode relations"

+++ app/database/migrations/2024_08_16_000000_create_anime_table.php
@@ -9,7 +9,7 @@ return new class extends Migration
     public function up(): void
     {
         Schema::create('anime', function (Blueprint $table) {
-            $table->unsignedSmallInteger('id')->primary();
+            $table->unsignedBigInteger('id')->primary();
             $table->string('title');
             $table->string('poster_url')->nullable();
             $table->string('type')->nullable();
